requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.2.0